moved get_node_by_path under a module
